refactor aria file names
