--- Integrations/ARIA/Pipfile
@@ -1,17 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-pylint = "*"
-pytest = "==5.0.1"
-pytest-mock = "*"
-requests-mock = "*"
-pytest-asyncio = "*"
-
-[packages]
-requests = "*"
-
-[requires]
-python_version = "3.7"
